Debug: Improve dry-run behavior

+++ app-common-io/src/main/java/eu/darken/sdmse/common/pkgs/pkgops/ipc/PkgOpsHost.kt
@@ -7,7 +7,9 @@ import android.content.pm.PackageManager
 import dagger.hilt.android.qualifiers.ApplicationContext
 import eu.darken.rxshell.cmd.Cmd
 import eu.darken.sdmse.common.debug.Bugs
+import eu.darken.sdmse.common.debug.Bugs.isDryRun
 import eu.darken.sdmse.common.debug.logging.Logging.Priority.ERROR
+import eu.darken.sdmse.common.debug.logging.Logging.Priority.INFO
 import eu.darken.sdmse.common.debug.logging.Logging.Priority.VERBOSE
 import eu.darken.sdmse.common.debug.logging.asLog
 import eu.darken.sdmse.common.debug.logging.log
@@ -22,6 +24,7 @@ import eu.darken.sdmse.common.pkgs.pkgops.LibcoreTool
 import eu.darken.sdmse.common.shell.SharedShell
 import eu.darken.sdmse.common.user.UserHandle2
 import kotlinx.coroutines.runBlocking
+import java.lang.Thread.sleep
 import javax.inject.Inject
 
 
@@ -85,25 +88,41 @@ class PkgOpsHost @Inject constructor(
         throw e.wrapToPropagate()
     }
 
-    override fun clearCacheAsUser(packageName: String, handleId: Int): Boolean = try {
-        log(TAG, VERBOSE) { "clearCache(packageName=$packageName, handleId=$handleId)..." }
-        runBlocking { pm.deleteApplicationCacheFilesAsUser(packageName, handleId) }
+    override fun clearCacheAsUser(packageName: String, handleId: Int, dryRun: Boolean): Boolean = try {
+        log(TAG, VERBOSE) { "clearCache(packageName=$packageName, handleId=$handleId, dryRun=$dryRun)..." }
+        if (dryRun) {
+            sleep(100)
+            true
+        } else {
+            runBlocking { pm.deleteApplicationCacheFilesAsUser(packageName, handleId) }
+        }
     } catch (e: Exception) {
         log(TAG, ERROR) { "clearCache(packageName=$packageName, handleId=$handleId) failed: ${e.asLog()}" }
         throw e.wrapToPropagate()
     }
 
-    override fun clearCache(packageName: String): Boolean = try {
-        log(TAG, VERBOSE) { "clearCache(packageName=$packageName)..." }
-        runBlocking { pm.deleteApplicationCacheFiles(packageName) }
+    override fun clearCache(packageName: String, dryRun: Boolean): Boolean = try {
+        log(TAG, VERBOSE) { "clearCache(packageName=$packageName, dryRun=$dryRun)..." }
+        if (dryRun) {
+            sleep(100)
+            true
+        } else {
+            runBlocking { pm.deleteApplicationCacheFiles(packageName) }
+        }
     } catch (e: Exception) {
         log(TAG, ERROR) { "clearCache(packageName=$packageName) failed: ${e.asLog()}" }
         throw e.wrapToPropagate()
     }
 
-    override fun trimCaches(desiredBytes: Long, storageId: String?): Boolean = try {
-        log(TAG, VERBOSE) { "trimCaches(desiredBytes=$desiredBytes, storageId=$storageId)..." }
-        runBlocking { pm.freeStorageAndNotify(desiredBytes, storageId) }
+    override fun trimCaches(desiredBytes: Long, storageId: String?, dryRun: Boolean): Boolean = try {
+        log(TAG, VERBOSE) { "trimCaches(desiredBytes=$desiredBytes, storageId=$storageId, dryRun=$dryRun)..." }
+        if (isDryRun) {
+            log(TAG, INFO) { "DRYRUN: not executing trimCaches($desiredBytes, $storageId)" }
+            sleep(2000)
+            true
+        } else {
+            runBlocking { pm.freeStorageAndNotify(desiredBytes, storageId) }
+        }
     } catch (e: Exception) {
         log(TAG, ERROR) { "trimCaches(desiredBytes=$desiredBytes, storageId=$storageId) failed: ${e.asLog()}" }
         throw e.wrapToPropagate()
